Swap out haikyuu ED theme
There is a conflict between the reddit post and the jury write-ups over which Haikyuu ED was nominated for 2016. I asked hamp (who was a 2016 ED juror) which one it was, and he said it was the S2 one. Reading the write-up it seems clear that it describes the S2 ED, and watching both EDs, I can see the S2 one getting nomed over the S3 one.
 
https://www.reddit.com/r/anime/comments/5saqbp/ranime_awards_2016_and_the_winners_are/

https://docs.google.com/document/d/1o15IbKZ_AocPxT1edYjIrsjTwTd7SG2VLqzot9Gq3Qg/edit#heading=h.kvsklwifddkh

https://animethemes.moe/video/HaikyuuS2-ED2-NCBD1080.webm

https://animethemes.moe/video/HaikyuuS3-ED1-NCBD1080.webm